+++ packages/phoenix-event-display/scripts/gen-api-docs.js
@@ -3,7 +3,7 @@ const { exec } = require('child_process');
 const path = require('path');
 
 if (!fs.existsSync('package.json')) {
-  console.log('Unable to run script.\nUse `npm run gen-api-docs` from directory containing package.json.');
+  console.log('Unable to run script.\nUse `yarn gen-api-docs` from directory containing package.json.');
   process.exit(1);
 }
 
@@ -13,7 +13,7 @@ if (!fs.existsSync('package.json')) {
 const DOCS_NAME = 'Phoenix API Docs';
 
 // Run from project directory containing package.json
-const COMPODOC_GEN = 'compodoc -p compodoc.conf.json -n \"' + DOCS_NAME + '\" -d ../../docs/api-docs';
+const COMPODOC_GEN = 'compodoc -p configs/compodoc.conf.json -n \"' + DOCS_NAME + '\" -d ../../docs/api-docs';
 
 console.log('Running command: %s', COMPODOC_GEN);